refacto+ use cleanenv for resolvebin
https://coveord.atlassian.net/browse/CDX-924
