+++ packages/cli-e2e/setup/utils.ts
@@ -154,9 +154,45 @@ export const resolveBinary = (programName: string) => {
   const spawner = spawnSync(whereOrWhich, [programName], {
     shell: true,
     encoding: 'utf-8',
+    env: getCleanEnv(),
   });
   return spawner.stdout.trim();
 };
 
+const registryEnv = process.env.E2E_USE_NPM_REGISTRY
+  ? {}
+  : {
+      npm_config_registry: 'http://localhost:4873',
+      YARN_NPM_REGISTRY_SERVER: 'http://localhost:4873',
+    };
+
+export function getCleanEnv(): Record<string, any> {
+  const env: Record<string, any> = {
+    ...process.env,
+    ...registryEnv,
+    npm_config_cache: process.env[npmCachePathEnvVar],
+  };
+  const excludeEnvVars = [
+    'npm_config_local_prefix',
+    'npm_package_json',
+    'INIT_CWD',
+  ];
+  const pathSep = process.platform === 'win32' ? ';' : ':';
+  const pathName = process.platform === 'win32' ? 'Path' : 'PATH';
+  const path = env[pathName].split(pathSep);
+  const filteredPath = path.filter(
+    (pathElement: string) => !isParent(env['GITHUB_WORKSPACE'], pathElement)
+  );
+  env[pathName] = filteredPath.join(pathSep);
+  for (const excludeVar of excludeEnvVars) {
+    delete env[excludeVar];
+  }
+  return env;
+}
+
 const appendCmdIfWindows = (cmd: TemplateStringsArray) =>
   `${cmd}${process.platform === 'win32' ? '.cmd' : ''}`;
+
+function isParent(parent: string, potentialChild: string) {
+  return resolve(potentialChild).startsWith(resolve(parent));
+}
