+++ packages/cli-e2e/utils/cli.ts
@@ -5,7 +5,7 @@ import {homedir} from 'os';
 import {ProcessManager} from './processManager';
 import {readJsonSync} from 'fs-extra';
 import {Terminal} from './terminal/terminal';
-import {npmCachePathEnvVar} from './npm';
+import {getCleanEnv} from '../setup/utils';
 
 export const isGenericYesNoPrompt = /\(y\/n\)[\s:]*$/i;
 
@@ -101,39 +101,4 @@ export function getConfig() {
   const pathToConfig = getConfigFilePath();
 
   return readJsonSync(pathToConfig);
-}
-
-export const registryEnv = process.env.E2E_USE_NPM_REGISTRY
-  ? {}
-  : {
-      npm_config_registry: 'http://localhost:4873',
-      YARN_NPM_REGISTRY_SERVER: 'http://localhost:4873',
-    };
-
-function getCleanEnv(): Record<string, any> {
-  const env: Record<string, any> = {
-    ...process.env,
-    ...registryEnv,
-    npm_config_cache: process.env[npmCachePathEnvVar],
-  };
-  const excludeEnvVars = [
-    'npm_config_local_prefix',
-    'npm_package_json',
-    'INIT_CWD',
-  ];
-  const pathSep = process.platform === 'win32' ? ';' : ':';
-  const pathName = process.platform === 'win32' ? 'Path' : 'PATH';
-  const path = env[pathName].split(pathSep);
-  const filteredPath = path.filter(
-    (pathElement: string) => !isParent(env['GITHUB_WORKSPACE'], pathElement)
-  );
-  env[pathName] = filteredPath.join(pathSep);
-  for (const excludeVar of excludeEnvVars) {
-    delete env[excludeVar];
-  }
-  return env;
-}
-
-function isParent(parent: string, potentialChild: string) {
-  return resolve(potentialChild).startsWith(resolve(parent));
 }
